rename view image modal to file preview modal and change imports
