+++ components/file_preview_modal/image_preview.jsx
