+++ components/file_preview_modal/index.js
@@ -8,7 +8,7 @@ import {getPost} from 'mattermost-redux/selectors/entities/posts';
 
 import {canDownloadFiles} from 'utils/file_utils.jsx';
 
-import ViewImage from './view_image.jsx';
+import FilePreviewModal from './file_preview_modal';
 
 function mapStateToProps(state, ownProps) {
     const config = getConfig(state);
@@ -21,4 +21,4 @@ function mapStateToProps(state, ownProps) {
     };
 }
 
-export default connect(mapStateToProps)(ViewImage);
+export default connect(mapStateToProps)(FilePreviewModal);
