+++ components/file_preview_modal/popover_bar/index.ts
