+++ components/file_preview_modal/popover_bar/popover_bar.tsx
